Assignment2/requirements.txt
djangorestframework~=3.16.0
Django~=5.2